:recycle: refactor: move data to VideoDetailsContent
move videodetails main video to VideoDetailsContent for reusability on FavoriteVideoDetais
