+++ src/pages/VideoDetails/VideoDetails.page.jsx
@@ -2,18 +2,10 @@ import React, { useContext, useEffect } from 'react';
 import { useParams } from 'react-router';
 import { useAuth0 } from '@auth0/auth0-react';
 import PreviewRelatedList from '../../components/PreviewRelatedList';
-import {
-  Heading,
-  OutlineButton,
-  Text,
-} from '../../theme/components/Foundation.component';
-import {
-  BodyContainer,
-  VideoContent,
-  VideoDataContainer,
-  VideoPlayer,
-} from './VideoDetails.styles';
+
+import { BodyContainer } from './VideoDetails.styles';
 import VideosContext from '../../providers/VideoProvider/VideoContext';
+import VideoDetailsContent from '../../components/VideoDetailsContent/VideoDetailsContent.component';
 
 function VideoDetails() {
   const { videoId } = useParams();
@@ -31,27 +23,18 @@ function VideoDetails() {
   const handleFavoriteVideo = () => {
     let favoriteList = localStorage.getItem('REACT-CHALLENGE-FAVORITE-VIDEOS') || [];
     favoriteList = JSON.parse(favoriteList);
-    console.log(favoriteList);
     favoriteList.push(videoData);
     localStorage.setItem('REACT-CHALLENGE-FAVORITE-VIDEOS', JSON.stringify(favoriteList));
   };
 
   return (
     <BodyContainer>
-      <VideoContent>
-        <VideoPlayer src={`https://www.youtube.com/embed/${videoId}`} />
-        {videoData && (
-          <VideoDataContainer>
-            <Heading>{videoData.snippet.title}</Heading>
-            <Text fontSize=".9rem">{videoData.snippet.description}</Text>
-            {isAuthenticated && (
-              <OutlineButton mt="1rem" onClick={handleFavoriteVideo}>
-                Add to favorites
-              </OutlineButton>
-            )}
-          </VideoDataContainer>
-        )}
-      </VideoContent>
+      <VideoDetailsContent
+        videoId={videoId}
+        videoData={videoData}
+        isAuthenticated={isAuthenticated}
+        handleFavoriteVideo={handleFavoriteVideo}
+      />
       {videos && <PreviewRelatedList videos={videos} />}
     </BodyContainer>
   );
